[Fix] LongTap
